Added redis connection handling
When initially connecting to docker, it will look for a special
environment variable called "REDIS_NAME" and use that as a clue
that it's safe to connect via 'redis'. This is due to the way docker
container linking works. If it isn't found, it simply tries to connect
to docker via localhost.