Removed extra babel dependencies
Instead of compiling es2015 -> es5, we instead transform only the needed
things that the v8 engine doesn't support (like desctructuring) with the
es2015-node5 package. This requires node5, but it's fine for us to use the
latest and greatest package. This should mean the code is slightly faster as
well, since we're using natively implemented features.