Implemented getMany and get-selected endpoint
Many times we'll want to do a transformation of ids, converting userIds
into actual user profiles and sending them back. Making a request for each one
is wasteful, so we can instead send an array of ids, and get an array of corresponding
user objects back. This lets us use userId references in other databases, but still
have access to the rich object if we need it by querying pylon.