Fixed mysterious socket hangup bug
The problem was the body parsing middleware was enabled globally for
the whole application. I'm not sure why it caused the error specifically, but
it must mess with the lower level response / request object provided by node.
When requests were forwarded without any body payload they worked as expected,
however as soon as body parameters were added the application would cease
to function because of a socket hang up error.

This is likely because the middleware authors never anticipated the use case
of forwarding requests from within an express application. By moving the middleware
to only the routers where it's needed, we can keep the body parsing functionality for
sections that need it (users) and not muck around with forwarding.