Custom Header Test
Forwarding custom headers is extremely simple. Specifying them in the options object
of the request will set them there, and they will be merged with any other headers that are
already set. This was actually tested with another service to confirm it could access the value.
This means we have the ability to send a user header and other useful contextual information
for microsevices.