Added docker-compose file
This makes it super simple to run pylon as service and get it running quickly.
Rather than running docker 3 times in a specific order and specifying 50 options on
the command line, the whole docker setup is specified in the YAML file, and the only
thing you need to do is make sure the environment variables are set.

Once they are, just `docker-compose up` and you've got a running pylon instance.