Add paths for serving static content
This basically allows us to have a static client website that uses interactive javascript,
completely separate from the API. Essentially they download the static assets, and from there
they make APIs entirely via ajax. Although to sign in, they go to a separate webpage since
that portion of the experience can't be automated as a result of oauth.